ABLASTR: IntervalsParser & SafeCast
Refactor and move out the `SafeCast` and `IntervalsParser` from
WarpX to ABLASTR.
