+++ Source/Utils/Parser/CMakeLists.txt
@@ -2,7 +2,7 @@ foreach(D IN LISTS WarpX_DIMS)
     warpx_set_suffix_dims(SD ${D})
     target_sources(lib_${SD}
       PRIVATE
-        IntervalsParser.cpp
+        BTDIntervalsParser.cpp
         ParserUtils.cpp
     )
 endforeach()
