+++ Source/Utils/Parser/ParserUtils.cpp
@@ -81,48 +81,6 @@ void utils::parser::get (const amrex::ParmParse& a_pp, std::string const& group,
     }
 }
 
-namespace {
-    template< typename int_type >
-    AMREX_FORCE_INLINE
-    int_type safeCastTo(const amrex::Real x, const std::string& real_name) {
-        auto result = int_type(0);
-        bool error_detected = false;
-        std::string assert_msg;
-        // (2.0*(numeric_limits<int>::max()/2+1)) converts numeric_limits<int>::max()+1 to a real ensuring accuracy to all digits
-        // This accepts x = 2**31-1 but rejects 2**31.
-        using namespace amrex::literals;
-        constexpr int_type half_max_plus_one = std::numeric_limits<int_type>::max()/2+1;
-        constexpr amrex::Real max_range = (2.0_rt*static_cast<amrex::Real>(half_max_plus_one));
-        if (x < max_range) {
-            if (std::ceil(x) >= std::numeric_limits<int_type>::min()) {
-                result = static_cast<int_type>(x);
-            } else {
-                error_detected = true;
-                assert_msg = "Negative overflow detected when casting " + real_name + " = " +
-                             std::to_string(x) + " to integer type";
-            }
-        } else if (x > 0) {
-            error_detected = true;
-            assert_msg =  "Overflow detected when casting " + real_name + " = " + std::to_string(x) + " to integer type";
-        } else {
-            error_detected = true;
-            assert_msg =  "NaN detected when casting " + real_name + " to integer type";
-        }
-        WARPX_ALWAYS_ASSERT_WITH_MESSAGE(!error_detected, assert_msg);
-        return result;
-   }
-}
-
-
-int utils::parser::safeCastToInt(const amrex::Real x, const std::string& real_name) {
-    return ::safeCastTo<int> (x, real_name);
-}
-
-
-long utils::parser::safeCastToLong(const amrex::Real x, const std::string& real_name) {
-    return ::safeCastTo<long> (x, real_name);
-}
-
 
 amrex::Parser utils::parser::makeParser (
     std::string const& parse_function, amrex::Vector<std::string> const& varnames)
