+++ Source/ablastr/utils/Make.package
@@ -1,4 +1,6 @@
 CEXE_sources += Communication.cpp
+CEXE_sources += IntervalsParser.cpp
+CEXE_sources += SafeCast.cpp
 CEXE_sources += SignalHandling.cpp
 CEXE_sources += TextMsg.cpp
 CEXE_sources += UsedInputsFile.cpp
